+++ projects/ngx-qgrid/src/lib/cell-handler/cell-handler.component.ts
@@ -109,9 +109,9 @@ export class CellHandlerComponent implements OnInit, AfterViewInit {
 						const target = domCell.element;
 						const scrollState = model.scroll();
 
-						const rect = table.view.rect('body-mid');
+						const headHeight = table.view.height('head-mid');
 
-						const top = Math.max(0, (target.offsetTop - scrollState.top));
+						const top = Math.max(headHeight, (target.offsetTop - scrollState.top));
 						const left = (target.offsetLeft - (cell.column.pin === 'mid' ? scrollState.left : 0));
 						const width = target.offsetWidth;
 						const height = target.offsetHeight;
@@ -121,12 +121,6 @@ export class CellHandlerComponent implements OnInit, AfterViewInit {
 							element.style.left = left + 'px';
 							element.style.width = width + 'px';
 							element.style.height = height + 'px';
-
-							console.log('rect.top ' + rect.top);
-							console.log('target.top ' + target.offsetTop);
-							console.log('scroll.top ' + scrollState.top);
-							console.log('top ' + top);
-							console.log('------------');
 						});
 					});