+++ projects/qgrid-core/mouse/mouse.state.d.ts
@@ -1,7 +1,10 @@
 import { Td } from '../dom/td';
 
+export declare type MouseStateCode = 'left' | 'right' | 'middle';
+export declare type MouseStateStatus = 'release' | 'down' | 'up';
+
 export declare class MouseState {
-	code: 'left' | 'right' | 'middle';
-	status: 'release' | 'down' | 'up';
+	code: MouseStateCode;
+	status: MouseStateStatus;
 	target: Td;
 }