+++ projects/qgrid-core/plugin/grid.plugin.d.ts
@@ -1,5 +1,4 @@
 import { Model } from '../model/model';
-import { ModelEvent } from '../model/model.event';
 import { Table } from '../dom/table';
 import { GridLet } from '../grid/grid.let';
 import { ObservableLike } from '../rx/rx';