+++ projects/qgrid-core/mouse/mouse.state.d.ts
@@ -1,10 +1,11 @@
 import { Td } from '../dom/td';
 
-export declare type MouseStateCode = 'left' | 'right' | 'middle';
+export declare type MouseStateCode = 'left' | 'right' | 'middle' | null;
 export declare type MouseStateStatus = 'release' | 'down' | 'up';
 
 export declare class MouseState {
 	code: MouseStateCode;
 	status: MouseStateStatus;
 	target: Td;
+	timestamp: number;
 }