+++ projects/qgrid-core/view/view.host.js
@@ -216,7 +216,8 @@ export class ViewHost {
 		model.mouse({
 			code: stringify(NO_BUTTON),
 			status: 'release',
-			target: null
+			target: null,
+			timestamp: Date.now(),
 		}, {
 			source: 'mouse.up'
 		});