+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [9.1.3] - 2020-05-29
+## [9.1.3] - 2020-06-01
 ### BREAKING
 * Remove q-grid-title, use `q-grid-caption` instead.